feat: 시연용 fcm 토큰 버튼 추가
